+++ packages/client/src/hooks/useNotification.ts
@@ -62,7 +62,9 @@ const useNotification = () => {
     if (permission === 'denied') {
       alert('Пожалуйста, разрешите уведомления в настройках вашего браузера')
     } else if (permission === 'granted') {
-      alert('Пожалуйста, отключите уведомления в настройках вашего браузера')
+      alert(
+        'Пожалуйста, сбросьте разрешение на уведомления в настройках вашего браузера'
+      )
     } else {
       requestPermission()
     }
